Move out types to live in its own island
