+++ packages/data/src/types/index.ts
@@ -1,0 +1,2 @@
+export * from './wp-data';
+export * from './rule-processor';
